requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
boto3==1.26.69
boto3-stubs==1.26.69
botocore==1.29.69
botocore-stubs==1.29.69
mypy-boto3-cloudformation==1.26.60
mypy-boto3-dynamodb==1.26.24
mypy-boto3-ec2==1.26.69
mypy-boto3-lambda==1.26.55
mypy-boto3-rds==1.26.47
mypy-boto3-s3==1.26.62
mypy-boto3-sqs==1.26.0.post1
CacheControl==0.12.11
cachetools==5.2.0
certifi>=2022.12.07
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
colorthief==0.2.1
cryptography>=39.0.1
Deprecated==1.2.13
discord.py==2.0.1
discord.py-pagination==0.0.5
exceptiongroup==1.0.0rc9
firebase-admin==6.0.1
frozenlist==1.3.3
gcloud==0.18.3
google-api-core==2.11.0
google-api-python-client==2.70.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.7.3
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
grpcio==1.51.1
grpcio-status==1.51.1
h11==0.14.0
httplib2==0.21.0
idna==3.4
jwcrypto==1.4.2
ldap3==2.9.1
lyricsgenius==3.0.1
msgpack==1.0.4
multidict==6.0.3
numpy==1.23.4
oauth2client==4.1.3
openai==0.27.0
outcome==1.2.0
pandas==1.5.1
Pillow==9.3.0
proto-plus==1.22.1
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.16.0
pydirectory==0.1.15
PyJWT==2.6.0
pyodbc==4.0.34
pyparsing==3.0.9
Pyrebase4==4.6.0
PySocks==1.7.1
python-dateutil==2.8.2
python-jwt==4.0.0
pytz==2022.5
redis>=4.4.4
requests==2.28.1
requests-toolbelt==0.10.1
rsa==4.9
selenium==4.5.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spotipy>=2.22.1
tabulate==0.9.0
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.9.2
uritemplate==4.1.1
urllib3==1.26.12
wrapt==1.14.1
wsproto==1.2.0
yarl==1.8.2
